Avoid model declaration redundancy (and thus potential DB mistakes)
