+++ test_call.py
@@ -5,10 +5,10 @@ import requests
 site = 'http://127.0.0.1:8008'
 
 allthems = requests.get(url='{0}/api/themes'.format(site))
-print allthems.json()
+print(allthems.json())
 
 quote = requests.get(url='{0}/api/random'.format(site))
-print quote.json()
+print(quote.json())
 
 nquote = requests.post(url='{0}/api/quote'.format(site), json={'theme': 'Absence'})
-print nquote.json()
+print(nquote.json())
